# various string search perf improvements