--- a/base/strings/search.jl
+++ b/base/strings/search.jl
@@ -1,12 +1,14 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
+@noinline __str_throw_boundserror(a, i) = throw(BoundsError(a, i))
+
 nothing_sentinel(i) = i == 0 ? nothing : i
 
 function findnext(pred::Fix2{<:Union{typeof(isequal),typeof(==)},<:AbstractChar},
                   s::String, i::Integer)
     if i < 1 || i > sizeof(s)
         i == sizeof(s) + 1 && return nothing
-        throw(BoundsError(s, i))
+        __str_throw_boundserror(s, i)
     end
     @inbounds isvalid(s, i) || string_index_err(s, i)
     c = pred.x
@@ -27,11 +29,11 @@ findnext(pred::Fix2{<:Union{typeof(isequal),typeof(==)},<:Union{Int8,UInt8}}, a:
 
 function _search(a::Union{String,ByteArray}, b::Union{Int8,UInt8}, i::Integer = 1)
     if i < 1
-        throw(BoundsError(a, i))
+         __str_throw_boundserror(a, i)
     end
     n = sizeof(a)
     if i > n
-        return i == n+1 ? 0 : throw(BoundsError(a, i))
+        return i == n+1 ? 0 :  __str_throw_boundserror(a, i)
     end
     p = pointer(a)
     q = ccall(:memchr, Ptr{UInt8}, (Ptr{UInt8}, Int32, Csize_t), p+i-1, b, n-i+1)
@@ -67,11 +69,11 @@ findprev(pred::Fix2{<:Union{typeof(isequal),typeof(==)},<:Union{Int8,UInt8}}, a:
 
 function _rsearch(a::Union{String,ByteArray}, b::Union{Int8,UInt8}, i::Integer = sizeof(a))
     if i < 1
-        return i == 0 ? 0 : throw(BoundsError(a, i))
+        return i == 0 ? 0 : __str_throw_boundserror(a, i)
     end
     n = sizeof(a)
     if i > n
-        return i == n+1 ? 0 : throw(BoundsError(a, i))
+        return i == n+1 ? 0 : __str_throw_boundserror(a, i)
     end
     p = pointer(a)
     q = ccall(:memrchr, Ptr{UInt8}, (Ptr{UInt8}, Int32, Csize_t), p, b, i)
@@ -107,7 +109,7 @@ findfirst(pattern::AbstractString, string::AbstractString) =
 # AbstractString implementation of the generic findnext interface
 function findnext(testf::Function, s::AbstractString, i::Integer)
     z = ncodeunits(s) + 1
-    1 ≤ i ≤ z || throw(BoundsError(s, i))
+    1 ≤ i ≤ z || __str_throw_boundserror(s, i)
     @inbounds i == z || isvalid(s, i) || string_index_err(s, i)
     for (j, d) in pairs(SubString(s, i))
         if testf(d)
@@ -124,7 +126,7 @@ function _searchindex(s::Union{AbstractString,ByteArray},
                       i::Integer)
     if isempty(t)
         return 1 <= i <= nextind(s,lastindex(s)) ? i :
-               throw(BoundsError(s, i))
+               __str_throw_boundserror(s, i)
     end
     t1, trest = Iterators.peel(t)
     while true
@@ -148,8 +150,8 @@ _nthbyte(s::String, i) = codeunit(s, i)
 _nthbyte(a::Union{AbstractVector{UInt8},AbstractVector{Int8}}, i) = a[i]
 
 function _searchindex(s::String, t::String, i::Integer)
-    # Check for fast case of a single byte
-    lastindex(t) == 1 && return something(findnext(isequal(t[1]), s, i), 0)
+    # Check for fast case of a single char
+    (ncodeunits(t) == 1 || lastindex(t) == 1) && return something(findnext(isequal(first(t)), s, i), 0)
     _searchindex(unsafe_wrap(Vector{UInt8},s), unsafe_wrap(Vector{UInt8},t), i)
 end
 
@@ -253,7 +255,15 @@ julia> findnext("Lang", "JuliaLang", 2)
 6:9
 ```
 """
-findnext(t::AbstractString, s::AbstractString, i::Integer) = _search(s, t, i)
+function findnext(t::AbstractString, s::AbstractString, i::Integer)
+    if t isa String && ncodeunits(t) == 1
+        # Optimization for the case of ascii single char `t`
+        v = nothing_sentinel(_search(s, first(t) % UInt8, i))
+        return v === nothing ? nothing : v:v
+    else
+        return _search(s, t, i)
+    end
+end
 
 """
     findlast(pattern::AbstractString, string::AbstractString)
@@ -270,17 +280,16 @@ julia> findfirst("Julia", "JuliaLang")
 1:5
 ```
 """
-findlast(pattern::AbstractString, string::AbstractString) =
-    findprev(pattern, string, lastindex(string))
+findlast(t::AbstractString, s::AbstractString) = findprev(t, s, lastindex(s))
 
 # AbstractString implementation of the generic findprev interface
 function findprev(testf::Function, s::AbstractString, i::Integer)
     if i < 1
-        return i == 0 ? nothing : throw(BoundsError(s, i))
+        return i == 0 ? nothing : __str_throw_boundserror(s, i)
     end
     n = ncodeunits(s)
     if i > n
-        return i == n+1 ? nothing : throw(BoundsError(s, i))
+        return i == n+1 ? nothing : __str_throw_boundserror(s, i)
     end
     # r[reverseind(r,i)] == reverse(r)[i] == s[i]
     # s[reverseind(s,j)] == reverse(s)[j] == r[j]
@@ -294,7 +303,7 @@ function _rsearchindex(s::AbstractString,
                        i::Integer)
     if isempty(t)
         return 1 <= i <= nextind(s, lastindex(s)) ? i :
-               throw(BoundsError(s, i))
+               __str_throw_boundserror(s, i)
     end
     t1, trest = Iterators.peel(Iterators.reverse(t))
     while true
@@ -426,7 +435,15 @@ julia> findprev("Julia", "JuliaLang", 6)
 1:5
 ```
 """
-findprev(t::AbstractString, s::AbstractString, i::Integer) = _rsearch(s, t, i)
+function findprev(t::AbstractString, s::AbstractString, i::Integer)
+     if t isa String && ncodeunits(t) == 1
+        # Optimization for the case of ascii single char `t`
+        v = nothing_sentinel(_rsearch(s, first(t) % UInt8, i))
+        return v === nothing ? nothing : v:v
+    else
+        _rsearch(s, t, i)
+    end
+end
 
 """
     occursin(needle::Union{AbstractString,Regex,AbstractChar}, haystack::AbstractString)