Before

```jl
julia> @btime findnext(isequal('v'), "abcdefghijklmnopqrstuv", 1)
  20.982 ns (0 allocations: 0 bytes)
22

julia> @btime findnext(",", "fsdfdsfdsfsdfds, sdfsfd", 1)
  39.258 ns (0 allocations: 0 bytes)
16:16

julia> @btime split("str,str,str,str,str,str,str", ",")
  662.151 ns (10 allocations: 432 bytes)

julia> @btime findlast(",", "fsdfds,fdsfsdfdssdfsfd")
  44.284 ns (0 allocations: 0 bytes)
7:7

julia> @btime findlast(isequal(','), "fsdfds,fdsfsdfdssdfsfd")
  25.692 ns (0 allocations: 0 bytes)
7
```

After:

```jl
julia> @btime findnext(isequal('v'), "abcdefghijklmnopqrstuv", 1)
  11.596 ns (0 allocations: 0 bytes)
22

julia> @btime findnext(",", "fsdfdsfdsfsdfds, sdfsfd", 1)
  19.166 ns (0 allocations: 0 bytes)
16:16

julia> @btime split("str,str,str,str,str,str,str", ",")
  406.869 ns (10 allocations: 432 bytes)

julia> @btime findlast(",", "fsdfds,fdsfsdfdssdfsfd")
  19.495 ns (0 allocations: 0 bytes)
7:7

julia> @btime findlast(isequal(','), "fsdfds,fdsfsdfdssdfsfd")
  22.811 ns (0 allocations: 0 bytes)
7
```

Fixes https://github.com/JuliaLang/julia/issues/29555