@KristofferC 2018-10-16
@nanosoldier `runbenchmarks("string", vs = ":master")`