@nanosoldier 2018-10-16
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29678#issuecomment-430405449) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/ca50061368ab1ba082b6f5fea8bdda81bf2b92c4/aaee133_vs_9cda6f9/report.md). cc @ararslan