@KristofferC 2018-10-16
@nanosoldier `runbenchmarks(ALL, vs = ":master")`