@nanosoldier 2018-10-17
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29678#issuecomment-430440789) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/2651947cbf944e9689802812ad36b151af3a5a15/3f3b1f6_vs_5d9dc65/report.md). cc @ararslan