@KristofferC 2021-04-20
Upon rethinking this, I don't think this fast path is really useful.